# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Tên API của bạn",
    "version": "1.0.0",
    "description": "Tài liệu mô tả API sử dụng Swagger UI",
    "contact": {
      "name": "Tên bạn hoặc nhóm phát triển",
      "email": "email@example.com"
    }
  },
  "servers": [
    {
      "url": "fe-api.training.ssit.company",
      "description": "Server testing"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateOrderRequest": {
        "type": "object",
        "required": [
          "items",
          "paymentMethod"
        ],
        "properties": {
          "customerName": {
            "type": "string",
            "maxLength": 255,
            "description": "Tên khách hàng",
            "example": "Nguyễn Văn A"
          },
          "customerPhone": {
            "type": "string",
            "pattern": "^(\\+84|84|0)(3|5|7|8|9)\\d{8}$",
            "description": "SĐT khách hàng (optional)",
            "example": "0987654321"
          },
          "items": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "object",
              "required": [
                "productId",
                "quantity",
                "unitPrice"
              ],
              "properties": {
                "productId": {
                  "type": "string",
                  "format": "uuid",
                  "description": "ID sản phẩm",
                  "example": "abc12345-e89b-12d3-a456-426614174003"
                },
                "quantity": {
                  "type": "integer",
                  "minimum": 1,
                  "description": "Số lượng",
                  "example": 2
                },
                "unitPrice": {
                  "type": "number",
                  "format": "decimal",
                  "minimum": 0,
                  "description": "Đơn giá",
                  "example": 50000
                }
              }
            }
          },
          "discountAmount": {
            "type": "number",
            "format": "decimal",
            "minimum": 0,
            "description": "Tiền giảm giá",
            "example": 15000,
            "default": 0
          },
          "paymentMethod": {
            "type": "string",
            "enum": [
              "CASH",
              "CARD",
              "TRANSFER"
            ],
            "description": "Phương thức thanh toán",
            "example": "CASH"
          },
          "notes": {
            "type": "string",
            "maxLength": 500,
            "description": "Ghi chú",
            "example": "Khách hàng VIP"
          }
        }
      },
...
      "OrdersListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Lấy danh sách đơn hàng thành công"
          },
          "data": {
            "type": "object",
            "properties": {
              "orders": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Order"
                }
              },
              "total": {
                "type": "integer",
                "example": 150
              },
              "page": {
                "type": "integer",
                "example": 1
              },
              "limit": {
                "type": "integer",
                "example": 20
              },
              "totalPages": {
                "type": "integer",
                "example": 8
              }
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/orders": {
      "get": {
        "summary": "Lấy danh sách đơn hàng",
        "description": "Lấy danh sách đơn hàng với phân trang và tìm kiếm",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Số trang"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Số lượng items mỗi trang"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Tìm kiếm theo số đơn hoặc tên KH"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "COMPLETED",
                "CANCELLED",
                "REFUNDED"
              ]
            },
            "description": "Lọc theo trạng thái"
          },
          {
            "in": "query",
            "name": "paymentMethod",
            "schema": {
              "type": "string",
              "enum": [
                "CASH",
                "CARD",
                "TRANSFER"
              ]
            },
            "description": "Lọc theo phương thức thanh toán"
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Từ ngày (YYYY-MM-DD)"
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Đến ngày (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách đơn hàng",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdersListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Chưa đăng nhập",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Tạo đơn hàng mới",
        "description": "Tạo đơn hàng mới từ giỏ hàng POS",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tạo đơn hàng thành công",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Order"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Chưa đăng nhập",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/stats": {
      "get": {
        "summary": "Lấy thống kê đơn hàng",
        "description": "Lấy thống kê tổng quan về đơn hàng và doanh thu",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thống kê đơn hàng",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderStatsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Chưa đăng nhập",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/reports/daily": {
      "get": {
        "summary": "Báo cáo doanh thu hàng ngày",
        "description": "Lấy báo cáo doanh thu và đơn hàng theo ngày",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Ngày báo cáo (YYYY-MM-DD), mặc định là hôm nay"
          }
        ],
        "responses": {
          "200": {
            "description": "Báo cáo doanh thu hàng ngày",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/DailySalesReport"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Chưa đăng nhập",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "summary": "Lấy thông tin đơn hàng theo ID",
        "description": "Lấy thông tin chi tiết của một đơn hàng dựa trên ID",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID của đơn hàng"
          }
        ],
        "responses": {
          "200": {
            "description": "Thông tin đơn hàng",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Order"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Chưa đăng nhập",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy đơn hàng",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/{id}/status": {
      "post": {
        "summary": "Cập nhật trạng thái đơn hàng",
        "description": "Cập nhật trạng thái đơn hàng (Admin có thể cập nhật mọi trạng thái, User chỉ có thể chuyển PENDING → COMPLETED)",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID của đơn hàng"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật trạng thái thành công",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Order"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Chưa đăng nhập",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền thực hiện",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy đơn hàng",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}